Read description please
I have added a hidden NSFW setting, for retrieving random images from a nsfw api. This can be activated by doing /nsfw on.
It has been completely disabled unless turned on. If off, the setting is invisible. If on, you will see

Other than that, bugfixes. For future reference, I have added many checks in here. They don't change the bot usage, but they keep the terminal a bit cleaner, which is what we want